p384

Struct Odd

#[repr(transparent)]
pub struct Odd<T>(T)
where
    T: ?Sized;

Wrapper type for odd integers.

These are frequently used in cryptography, e.g. as a modulus.

Tuple Fields
0: T
Implementations

impl<const LIMBS: usize> Odd<Int<LIMBS>>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Odd<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Odd<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd( &self, rhs: &NonZero<Int<LIMBS>>, ) -> XgcdOutput<LIMBS, Odd<Uint<LIMBS>>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl<T> Odd<T>

pub fn new(n: T) -> CtOption<Odd<T>>
where T: Integer,

Create a new odd integer.

pub fn get(self) -> T

Returns the inner value.

pub const fn get_copy(self) -> T
where T: Copy,

Returns a copy of the inner value for Copy types.

This allows the function to be const fn, since Copy is implicitly !Drop, which avoids problems around const fn destructors.

pub fn into_nz(self) -> NonZero<T>

All odd integers are definitionally non-zero, so we can convert into the equivalent NonZero type.

impl<T> Odd<T>
where T: ?Sized,

pub const fn as_ref(&self) -> &T

Provides access to the contents of Odd in a const context.

pub const fn as_nz_ref(&self) -> &NonZero<T>

All odd integers are definitionally non-zero, so we can also obtain a reference to the equivalent NonZero type.

impl<T> Odd<T>
where T: Bounded + ?Sized,

pub const BITS: u32 = T::BITS

Total size of the represented integer in bits.

pub const BYTES: usize = T::BYTES

Total size of the represented integer in bytes.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub const fn from_be_hex(hex: &str) -> Odd<Uint<LIMBS>>

Create a new OddUint from the provided big endian hex string.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

if the value is even.

pub const fn from_le_hex(hex: &str) -> Odd<Uint<LIMBS>>

Create a new Odd<Uint<LIMBS>> from the provided little endian hex string.

Panics

if the hex is malformed or not zero-padded accordingly for the size.

if the value is even.

pub const fn as_uint_ref(&self) -> &Odd<UintRef>

Borrow this OddUint as a &OddUintRef.

pub const fn resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl<const LIMBS: usize> Odd<Int<LIMBS>>

pub const fn abs_sign(&self) -> (Odd<Uint<LIMBS>>, Choice)

The sign and magnitude of this Odd<Int<{LIMBS}>>.

pub const fn abs(&self) -> Odd<Uint<LIMBS>>

The magnitude of this Odd<Int<{LIMBS}>>.

impl Odd<UintRef>

pub const fn to_uint_resize<const T: usize>(&self) -> Odd<Uint<T>>

Construct an Odd<Uint<T>> from the unsigned integer value, truncating the upper bits if the value is too large to be represented.

impl Odd<BoxedUint>

pub const fn as_uint_ref(&self) -> &Odd<UintRef>

Borrow this OddBoxedUint as a &OddUintRef.

pub fn random<R>(rng: &mut R, bit_length: u32) -> Odd<BoxedUint>
where R: TryRng + ?Sized,

Generate a random Odd<Uint<T>>.

impl<const LIMBS: usize> Odd<Uint<LIMBS>>

pub const fn gcd_unsigned(&self, rhs: &Uint<LIMBS>) -> Odd<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

pub const fn gcd_unsigned_vartime(&self, rhs: &Uint<LIMBS>) -> Odd<Uint<LIMBS>>

Compute the greatest common divisor of self and rhs.

Executes in variable time w.r.t. all input parameters.

pub const fn xgcd( &self, rhs: &Odd<Uint<LIMBS>>, ) -> XgcdOutput<LIMBS, Odd<Uint<LIMBS>>>

Execute the Extended GCD algorithm.

Given (self, rhs), computes (g, x, y) s.t. self * x + rhs * y = g = gcd(self, rhs).

impl Odd<UintRef>

pub const fn invert_mod_u64(&self) -> u64

Returns the multiplicative inverse of the argument modulo 2^64.