p384

Struct Scalar

pub struct Scalar(/* private fields */);

Scalars are elements in the finite field modulo n.

Trait impls

Much of the important functionality of scalars is provided by traits from the ff crate, which is re-exported as p384::elliptic_curve::ff:

Field - represents elements of finite fields and provides:

Field::random - generate a random scalar

double, square, and invert operations

Bounds for Add, Sub, Mul, and Neg (as well as *Assign equivalents)

Bounds for ConditionallySelectable from the subtle crate

PrimeField - represents elements of prime fields and provides:

from_repr/to_repr for converting field elements from/to big integers.

multiplicative_generator and root_of_unity constants.

PrimeFieldBits - operations over field elements represented as bits (requires bits feature)

Please see the documentation for the relevant traits for more information.

serde support

When the serde feature of this crate is enabled, the Serialize and Deserialize traits are impl’d for this type.

The serialization is a fixed-width big endian encoding. When used with textual formats, the binary data is encoded as hexadecimal.

Implementations

impl Scalar

pub const ZERO: Self

Zero element.

pub const ONE: Self

Multiplicative identity.

pub fn from_bytes(repr: &FieldBytes) -> CtOption<Self>

Create a Scalar from a canonical big-endian representation.

pub fn from_slice(slice: &[u8]) -> Result<Self>

Decode Scalar from a big endian byte slice.

pub fn from_uint(uint: U384) -> CtOption<Self>

Decode Scalar from U384 converting it into Montgomery form:

w * R^2 * R^-1 mod p = wR mod p

pub const fn from_u64(w: u64) -> Self

Convert a u64 into a Scalar.

pub fn to_bytes(self) -> FieldBytes

Returns the big-endian encoding of this Scalar.

pub const fn to_canonical(self) -> U384

Translate Scalar out of the Montgomery domain, returning a U384 in canonical form.

pub fn is_odd(&self) -> Choice

Determine if this Scalar is odd in the SEC1 sense: self mod 2 == 1.

Returns

If odd, return Choice(1). Otherwise, return Choice(0).

pub fn is_even(&self) -> Choice

Determine if this Scalar is even in the SEC1 sense: self mod 2 == 0.

Returns

If even, return Choice(1). Otherwise, return Choice(0).

pub fn is_zero(&self) -> Choice

Determine if this Scalar is zero.

Returns

If zero, return Choice(1). Otherwise, return Choice(0).

pub const fn add(&self, rhs: &Self) -> Self

Add elements.

pub const fn double(&self) -> Self

Double element (add it to itself).

pub const fn sub(&self, rhs: &Self) -> Self

Subtract elements.

pub const fn multiply(&self, rhs: &Self) -> Self

Multiply elements.

pub const fn neg(&self) -> Self

Negate element.

pub const fn square(&self) -> Self

Compute modular square.

pub const fn pow_vartime(&self, exp: &[u64]) -> Self

Returns self^exp, where exp is a little-endian integer exponent.

This operation is variable time with respect to the exponent.

If the exponent is fixed, this operation is effectively constant time.

impl Scalar

pub fn invert(&self) -> CtOption<Self>

Compute Scalar inversion: 1 / self.

pub fn sqrt(&self) -> CtOption<Self>

Compute modular square root.

pub const fn shr_vartime(&self, shift: usize) -> Scalar

Right shifts the scalar.

Note: not constant-time with respect to the shift parameter.